Deepak - Bug Fix
Fixed a small bug in image-mask blending imseriesmaskshowrgb.m ... 

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@9132 0bcb941b-74d1-0310-bb26-e6636707f451

+++ graphics/imageSeriesVisualization/imseriesmaskshowrgb.m
@@ -359,7 +359,7 @@ function [ rgbSlice ] = getRGBSlice(data)
             else
                 rgbMaskSlice = squeeze(data.masks(i).im(:,:,data.sliceno,:));
             end
-            rgbSlice = uint8( double((1 - data.alpha(i)) * rgbSlice) + 255 * data.alpha(i) * rgbMaskSlice );
+            rgbSlice = uint8( double((1 - data.alpha(i)) * double(rgbSlice)) + 255.0 * data.alpha(i) * double(rgbMaskSlice) );
         end
     end
